Epic Commit... Continuing to work on Ticket Queues
Decided on a HABTM relationship between users and ticket queues. Figure this way, you could have group queues... might be dumb but I want it so kiss off....

setup the HABTM relationship in the models, and added the join table. specified the join table name cause I wasn't sure if I did the naming exactly right.

Filled out the ticket queue controller a bit and setup some views outside of the testing specs to try to make this work...